// src/lib/api/generated/model/createdVia.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type CreatedVia = typeof CreatedVia[keyof typeof CreatedVia];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreatedVia = {
  client_self: 'client_self',
  professional: 'professional',
} as const;